lib: cleanup, const-correctness, and -Werror
Squashed commit of the following:

commit 9e25b3f7e823018c57b8c198a6821318de2a3266
Author:     Vassili Tchersky <vt+git@vbcy.org>
AuthorDate: Wed Aug 26 10:32:24 2026 +0000
Commit:     Vassili Tchersky <vt+git@vbcy.org>
CommitDate: Wed Aug 26 23:20:02 2026 +0000

    lib: clean the tree for Clang under more warnings and -Werror

    A bunch of diagnostics were emited by Clang. A lot of casts to
    incompatible function types, used a lot by callbacks and checksums
    interfaces. All those theoritically potential UBs are gone.

    Also, rm_traverse_directory() returned void, but its return value was
    used as a boolean to decide if rm_mds_factory() should wait 100 μs.

commit f5072a36ed04653b8ebe330f785b25acb1ee67c4
Author:     Cebtenzzre <cebtenzzre@gmail.com>
AuthorDate: Sat Sep 25 17:33:03 2021 -0400
Commit:     Ivan Shapovalov <intelfx@intelfx.name>
CommitDate: Fri Aug 21 17:02:20 2026 +0300

    -Werror, more warnings, and warning fixes

+++ lib/formats/stats.c
@@ -96,7 +96,7 @@ static void rm_fmt_prog(RmSession *session,
     fprintf(out, _("%s%15ld%s Other lint items\n"), MAYBE_RED(out, session),
             (long)session->other_lint_cnt, MAYBE_RESET(out, session));
 
-    gfloat elapsed = g_timer_elapsed(session->timer_since_proc_start, NULL);
+    gdouble elapsed = g_timer_elapsed(session->timer_since_proc_start, NULL);
 
     char *elapsed_time = rm_format_elapsed_time(elapsed, 5);
     fprintf(out, _("%s%15s%s of time spent scanning\n"), MAYBE_RED(out, session),
