+++ lib/formats/summary.c
@@ -107,7 +107,7 @@ static void rm_fmt_prog(RmSession *session,
                       _("This run was a replay from a previous run. No I/O done!\n"));
     }
 
-    gfloat elapsed = g_timer_elapsed(session->timer_since_proc_start, NULL);
+    gdouble elapsed = g_timer_elapsed(session->timer_since_proc_start, NULL);
     char *elapsed_time = rm_format_elapsed_time(elapsed, 3);
     ARROW fprintf(out, _("Scanning took in total %s%s%s.\n"), MAYBE_RED(out, session),
                   elapsed_time, MAYBE_RESET(out, session));
